+++ libraries/base/Control/Monad.hs
@@ -335,7 +335,7 @@ infixl 4 <$!>
 
 -- | Strict version of 'Data.Functor.<$>'.
 --
--- /Since: 4.7.1.0/
+-- /Since: 4.8.0.0/
 (<$!>) :: Monad m => (a -> b) -> m a -> m b
 {-# INLINE (<$!>) #-}
 f <$!> m = do