+++ libraries/base/changelog.md
@@ -18,6 +18,8 @@
     enabled, so that the `Monoid` instance for `Proxy` are polykinded
     like `Proxy` itself is.
 
+  * Make `abs` handle (-0.0) correctly per IEEE-754.
+
 ## 4.7.0.1  *Jul 2014*
 
   * Bundled with GHC 7.8.3