+++ libraries/base/changelog.md
@@ -1,8 +1,10 @@
 # Changelog for [`base` package](http://hackage.haskell.org/package/base)
 
-## 4.7.1.0 *TBA*
+## 4.8.0.0  *TBA*
 
   * Bundled with GHC 7.10.1
+
+  * Make `Applicative` a superclass of `Monad`
 
   * Add reverse application operator `Data.Function.(&)`